@@ -1,34 +1,33 @@
 {
     "name": "moravianlibrary/ziskej-api-php-client",
-    "version": "7.0.3",
+    "version": "8.0.0",
     "description": "Ziskej API PHP client",
     "type": "library",
     "license": "MIT",
     "require": {
-        "php": "^7.4 || ^8",
+        "php": "^8.1",
         "ext-json": "*",
-        "consistence-community/consistence": "^2.1",
-        "php-http/client-common": "^2.0",
-        "php-http/httplug": "^2.0",
-        "php-http/message-factory": "^1.0",
-        "php-http/discovery": "^1.0",
-        "php-http/logger-plugin": "^1.1",
-        "psr/http-message": "^1.0",
-        "smartemailing/types": "^v1.2.0"
+        "php-http/client-common": "^2.7",
+        "php-http/httplug": "^2.4",
+        "php-http/message-factory": "^1.1",
+        "php-http/discovery": "^1.19",
+        "php-http/logger-plugin": "^1.3",
+        "psr/http-message": "^1.1",
+        "smartemailing/types": "^v1.2"
     },
     "require-dev": {
-        "devcoder-xyz/php-dotenv": "^1.0",
-        "graylog2/gelf-php": "^1.6",
-        "guzzlehttp/psr7": "^2.0",
-        "lcobucci/jwt": "^4.0",
-        "monolog/monolog": "^1.24",
-        "nunomaduro/phpinsights": "^2.6",
+        "devcoder-xyz/php-dotenv": "^1.1",
+        "graylog2/gelf-php": "^1.7",
+        "guzzlehttp/psr7": "^2.5",
+        "lcobucci/jwt": "^4.3",
+        "monolog/monolog": "^1.27",
+        "nunomaduro/phpinsights": "^2.8",
         "php-http/guzzle7-adapter": "^1.0",
-        "php-http/mock-client": "^1.0",
-        "php-http/message": "^1.0",
-        "phpstan/phpstan": "^1.8",
-        "phpunit/phpunit": "^9.5",
-        "squizlabs/php_codesniffer": "^3.4"
+        "php-http/mock-client": "^1.6",
+        "php-http/message": "^1.16",
+        "phpstan/phpstan": "^1.10",
+        "phpunit/phpunit": "^9.6",
+        "squizlabs/php_codesniffer": "^3.7"
     },
     "autoload": {
         "psr-4": {
@@ -47,13 +46,12 @@
     "scripts": {
         "phpcbf": "phpcbf --standard=PSR12 --extensions=php --encoding=utf-8 --tab-width=4 -spv src tests",
         "phpcs": "phpcs --standard=PSR12 --extensions=php  --encoding=utf-8 --tab-width=4 -spv src tests",
-        "phpstan": "phpstan analyse -l 7 --memory-limit=500M --debug --xdebug -c phpstan.neon",
+        "phpstan": "phpstan analyse -l 7 --memory-limit=500M --debug  -c phpstan.neon",
         "phpunit": "phpunit --colors=always -c phpunit.xml",
         "test": [
             "@phpcbf",
             "@phpcs",
-            "@phpstan",
-            "@phpunit"
+            "@phpstan"
         ]
     },
     "config": {
