@@ -14,7 +14,9 @@ use Mzk\ZiskejApi\ResponseModel\Library;
 use Mzk\ZiskejApi\ResponseModel\LibraryCollection;
 use Mzk\ZiskejApi\ResponseModel\MessageCollection;
 use Mzk\ZiskejApi\ResponseModel\Ticket;
-use Mzk\ZiskejApi\ResponseModel\TicketsCollection;
+use Mzk\ZiskejApi\ResponseModel\TicketCollection;
+use Mzk\ZiskejApi\ResponseModel\TicketEdd;
+use Mzk\ZiskejApi\ResponseModel\TicketMvs;
 
 final class Api
 {
@@ -65,46 +67,8 @@ final class Api
                 '/libraries',
                 [],
                 [
-                    'service' => LibraryServiceType::ANY_ZK,
+                    'service' => LibraryServiceType::ANY_ZK->value,
                     'include_deactivated' => 1,
-                ]
-            )
-        );
-
-        switch ($apiResponse->getStatusCode()) {
-            case 200:
-                $contents = $apiResponse->getBody()->getContents();
-                $array = json_decode($contents, true);
-                if (isset($array['items']) && is_array($array['items'])) {
-                    return LibraryCollection::fromArray($array['items']);
-                }
-                return new LibraryCollection();
-
-            default:
-                throw new ApiResponseException($apiResponse);
-        }
-    }
-
-    /**
-     * List all active libraries
-     * GET /libraries
-     *
-     * @return \Mzk\ZiskejApi\ResponseModel\LibraryCollection
-     *
-     * @throws \Mzk\ZiskejApi\Exception\ApiResponseException
-     * @throws \Psr\Http\Client\ClientExceptionInterface
-     *
-     * @deprecated use Api::getLibrariesMvsActive() or Api::getLibrariesEddActive
-     */
-    public function getLibrariesActive(): LibraryCollection
-    {
-        $apiResponse = $this->apiClient->sendApiRequest(
-            new ApiRequest(
-                'GET',
-                '/libraries',
-                [],
-                [
-                    'service' => LibraryServiceType::MVS_ZK,
                 ]
             )
         );
@@ -140,7 +104,7 @@ final class Api
                 '/libraries',
                 [],
                 [
-                    'service' => LibraryServiceType::MVS_ZK,
+                    'service' => LibraryServiceType::MVS_ZK->value,
                 ]
             )
         );
@@ -176,7 +140,7 @@ final class Api
                 '/libraries',
                 [],
                 [
-                    'service' => LibraryServiceType::EDD_ZK,
+                    'service' => LibraryServiceType::EDD_ZK->value,
                 ]
             )
         );
@@ -338,13 +302,13 @@ final class Api
      *
      * @param string $eppn
      *
-     * @return \Mzk\ZiskejApi\ResponseModel\TicketsCollection List of tickets with details
+     * @return \Mzk\ZiskejApi\ResponseModel\TicketCollection List of tickets with details
      *
      * @throws \Mzk\ZiskejApi\Exception\ApiResponseException
      * @throws \Psr\Http\Client\ClientExceptionInterface
      * @throws \Exception
      */
-    public function getTickets(string $eppn): TicketsCollection
+    public function getTickets(string $eppn): TicketCollection
     {
         $apiResponse = $this->apiClient->sendApiRequest(
             new ApiRequest(
@@ -366,9 +330,9 @@ final class Api
                 $array = json_decode($contents, true);
 
                 if (isset($array['items']) && is_array($array['items'])) {
-                    $tickets = TicketsCollection::fromArray($array['items']);
+                    $tickets = TicketCollection::fromArray($array['items']);
                 } else {
-                    $tickets = new TicketsCollection();
+                    $tickets = new TicketCollection();
                 }
                 break;
             default:
@@ -384,13 +348,13 @@ final class Api
      *
      * @param string $eppn
      *
-     * @return \Mzk\ZiskejApi\ResponseModel\TicketsCollection List of tickets with details
+     * @return \Mzk\ZiskejApi\ResponseModel\TicketCollection List of tickets with details
      *
      * @throws \Mzk\ZiskejApi\Exception\ApiResponseException
      * @throws \Psr\Http\Client\ClientExceptionInterface
      * @throws \Exception
      */
-    public function getTicketsMvs(string $eppn): TicketsCollection
+    public function getTicketsMvs(string $eppn): TicketCollection
     {
         $apiResponse = $this->apiClient->sendApiRequest(
             new ApiRequest(
@@ -400,7 +364,7 @@ final class Api
                     ':eppn' => $eppn,
                 ],
                 [
-                    'ticket_type' => TicketType::MVS,
+                    'ticket_type' => TicketType::MVS->value,
                     'expand' => 'detail',
                     'include_closed' => 1,
                 ]
@@ -413,9 +377,9 @@ final class Api
                 $array = json_decode($contents, true);
 
                 if (isset($array['items']) && is_array($array['items'])) {
-                    $tickets = TicketsCollection::fromArray($array['items']);
+                    $tickets = TicketCollection::fromArray($array['items']);
                 } else {
-                    $tickets = new TicketsCollection();
+                    $tickets = new TicketCollection();
                 }
                 break;
             default:
@@ -431,13 +395,13 @@ final class Api
      *
      * @param string $eppn
      *
-     * @return \Mzk\ZiskejApi\ResponseModel\TicketsCollection List of tickets with details
+     * @return \Mzk\ZiskejApi\ResponseModel\TicketCollection List of tickets with details
      *
      * @throws \Mzk\ZiskejApi\Exception\ApiResponseException
      * @throws \Psr\Http\Client\ClientExceptionInterface
      * @throws \Exception
      */
-    public function getTicketsEdd(string $eppn): TicketsCollection
+    public function getTicketsEdd(string $eppn): TicketCollection
     {
         $apiResponse = $this->apiClient->sendApiRequest(
             new ApiRequest(
@@ -447,7 +411,7 @@ final class Api
                     ':eppn' => $eppn,
                 ],
                 [
-                    'ticket_type' => TicketType::EDD,
+                    'ticket_type' => TicketType::EDD->value,
                     'expand' => 'detail',
                     'include_closed' => 1,
                 ]
@@ -460,9 +424,9 @@ final class Api
                 $array = json_decode($contents, true);
 
                 if (isset($array['items']) && is_array($array['items'])) {
-                    $tickets = TicketsCollection::fromArray($array['items']);
+                    $tickets = TicketCollection::fromArray($array['items']);
                 } else {
-                    $tickets = new TicketsCollection();
+                    $tickets = new TicketCollection();
                 }
                 break;
             default:
@@ -506,7 +470,7 @@ final class Api
 
                 if (!isset($array['id'])) {
                     throw new ApiException(
-                        'Ziskej API error: API did not return "id" parameter.'  //@todo jina zprava
+                        'Ziskej API error: Ziskej API response "id" parameter is missing.'
                     );
                 }
 
@@ -545,7 +509,12 @@ final class Api
         switch ($apiResponse->getStatusCode()) {
             case 200:
                 $contents = $apiResponse->getBody()->getContents();
-                return Ticket::fromArray(json_decode($contents, true));
+                $array = json_decode($contents, true);
+                return match ($array['ticket_type']) {
+                    TicketType::MVS->value => TicketMvs::fromArray($array),
+                    TicketType::EDD->value => TicketEdd::fromArray($array),
+                    default => null,
+                };
             case 404:
                 return null;
             default:
@@ -684,14 +653,14 @@ final class Api
     {
         $apiResponse = $this->apiClient->sendApiRequest(
             new ApiRequest(
-                'PUT',
-                '/readers/:eppn/tickets/:ticket_id/messages',
-                [
+                method: 'PUT',
+                endpoint: '/readers/:eppn/tickets/:ticket_id/messages',
+                urlQuery: [
                     ':eppn' => $eppn,
                     ':ticket_id' => $ticketId,
                 ],
-                [],
-                $messages->toArray()
+                paramsUrl: [],
+                paramsData: $messages->toArray()
             )
         );
 
@@ -706,26 +675,23 @@ final class Api
     /**
      * Get EDD fee estimate
      *
-     * @param int $number_of_pages
-     * @param string $edd_subtype
+     * @param int $numberOfPages
+     * @param \Mzk\ZiskejApi\Enum\TicketEddSubtype $eddSubtype
      *
      * @return \Mzk\ZiskejApi\ResponseModel\EddEstimate
      *
-     * @throws \Consistence\Enum\InvalidEnumValueException
      * @throws \Mzk\ZiskejApi\Exception\ApiResponseException
      * @throws \Psr\Http\Client\ClientExceptionInterface
      */
-    public function getEddEstimateFee(int $number_of_pages, string $edd_subtype): EddEstimate
+    public function getEddEstimateFee(int $numberOfPages, TicketEddSubtype $eddSubtype): EddEstimate
     {
-        TicketEddSubtype::checkValue($edd_subtype);
-
         $apiRequest = new ApiRequest(
             'GET',
             '/service/edd/estimate',
             [],
             [
-                'number_of_pages' => $number_of_pages,
-                'edd_subtype' => $edd_subtype,
+                'number_of_pages' => $numberOfPages,
+                'edd_subtype' => $eddSubtype->value,
             ]
         );
 
