@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Mzk\ZiskejApi\Enum;
 
-abstract class TicketEddSubtype extends BaseEnum
+enum TicketEddSubtype: string
 {
-    public const ARTICLE = 'article';
-    public const SELECTION = 'selection';
+    case ARTICLE = 'article';
+    case SELECTION = 'selection';
 }
