@@ -4,8 +4,8 @@ declare(strict_types=1);
 
 namespace Mzk\ZiskejApi\Enum;
 
-abstract class TicketType extends BaseEnum
+enum TicketType: string
 {
-    public const MVS = 'mvs';
-    public const EDD = 'edd';
+    case MVS = 'mvs';
+    case EDD = 'edd';
 }
