@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Mzk\ZiskejApi\Enum;
 
-abstract class ZiskejSettings extends BaseEnum
+abstract class ZiskejSettings
 {
     public const EDD_SELECTION_MAX_PAGES = 20;
 }
