@@ -4,9 +4,8 @@ declare(strict_types=1);
 
 namespace Mzk\ZiskejApi\RequestModel;
 
-use Consistence\InvalidArgumentTypeException;
 use DateTimeImmutable;
-use Mzk\ZiskejApi\Enum\TicketEddDocDataSource;
+use Mzk\ZiskejApi\Enum\TicketDataSource;
 use Mzk\ZiskejApi\Enum\TicketEddSubtype;
 use Mzk\ZiskejApi\Enum\TicketType;
 
@@ -14,195 +13,55 @@ final class TicketEddRequest extends TicketRequest
 {
     /**
      * Ticket type
-     *
-     * @see \Mzk\ZiskejApi\Enum\TicketType
      */
     private const TICKET_TYPE = TicketType::EDD;
 
     /**
-     * Typ vytvoření objednávky (automatická, manuální)
-     *
-     * @var string
-     *
-     * @see \Mzk\ZiskejApi\Enum\TicketEddDocDataSource
-     */
-    protected string $ticketDocDataSource;
-
-    /**
-     * article (pro článek), selection (pro výňatek z monografie a periodika)
-     *
-     * @var string
-     *
-     * @see \Mzk\ZiskejApi\Enum\TicketEddSubtype
-     */
-    protected string $eddSubtype;
-
-    /**
-     * Název časopisu
-     *
-     * @var string
-     */
-    protected string $docTitleIn;
-
-    /**
-     * Název článku
-     *
-     * @var string
-     */
-    protected string $docTitle;
-
-    /**
-     * Document ID
-     *
-     * @var string
-     */
-    protected string $documentId;
-
-    /**
-     * Alternative document IDs
-     *
-     * @var array<string>
-     */
-    protected array $documentAltIds = [];
-
-    /**
-     * Parent document ID
-     *
-     * @var string|null
-     */
-    protected ?string $docIdIn = null;
-
-    /**
-     * Reader's note to librarian
-     *
-     * @var string|null
-     */
-    protected ?string $readerNote = null;
-
-    /**
-     * Rok
-     *
-     * @var string|null
-     */
-    protected ?string $docNumberYear = null;
-
-    /**
-     * Ročník
-     *
-     * @var string|null
-     */
-    protected ?string $docNumberPyear = null;
-
-    /**
-     * Číslo
-     *
-     * @var string|null
-     */
-    protected ?string $docNumberPnumber = null;
-
-    /**
-     * Svazek
-     *
-     * @var string|null
-     */
-    protected ?string $docVolume = null;
-
-    /**
-     * Rozsah stran, číslo první požadované strany
-     *
-     * @var int|null
-     */
-    protected ?int $pagesFrom = null;
-
-    /**
-     * Rozsah stran, číslo poslední požadované strany
-     *
-     * @var int|null
-     */
-    protected ?int $pagesTo = null;
-
-    /**
-     * Autor
-     *
-     * @var string|null
-     */
-    protected ?string $docAuthor = null;
-
-    /**
-     * Místo a rok vydání
-     *
-     * @var string|null
-     */
-    protected ?string $docIssuer = null;
-
-    /**
-     * ISSN
-     *
-     * @var string|null
-     */
-    protected ?string $docISSN = null;
-
-    /**
-     * ISBN
-     *
-     * @var string|null
-     */
-    protected ?string $docISBN = null;
-
-    /**
-     * Citace
-     *
-     * @var string|null
-     */
-    protected ?string $docCitation = null;
-
-    /**
-     * Poznámka k objednávce
-     *
-     * @var string|null
-     */
-    protected ?string $docNote = null;
-
-    /**
-     * Requested date
-     *
-     * @var \DateTimeImmutable|null
-     */
-    protected ?DateTimeImmutable $dateRequested = null;
-
-    /**
-     * Ticket EDD constructor.
-     *
-     * @param string $ticketDocDataSource
-     * @param string $eddSubtype
-     * @param string $docTitleIn
-     * @param string $docTitle
-     * @param string|null $documentId
-     *
-     * @throws \Consistence\Enum\InvalidEnumValueException
-     * @throws \Consistence\InvalidArgumentTypeException
+     * @param TicketDataSource $ticketDocDataSource Typ vytvoření objednávky (automatická, manuální)
+     * @param TicketEddSubtype $eddSubtype article (pro článek), selection (pro výňatek z monografie a periodika)
+     * @param string $docTitleIn Název časopisu
+     * @param string $docTitle Název článku
+     * @param string|null $documentId Document ID
+     * @param ?array<string> $documentAltIds Alternative document IDs
+     * @param string|null $docIdIn Parent document ID
+     * @param string|null $readerNote Reader's note to librarian
+     * @param string|null $docNumberYear Rok
+     * @param string|null $docNumberPyear Ročník
+     * @param string|null $docNumberPnumber Číslo
+     * @param string|null $docVolume Svazek
+     * @param int|null $pagesFrom Rozsah stran, číslo první požadované strany
+     * @param int|null $pagesTo Rozsah stran, číslo poslední požadované strany
+     * @param string|null $docAuthor Autor
+     * @param string|null $docIssuer Místo a rok vydání
+     * @param string|null $docISSN ISSN
+     * @param string|null $docISBN ISBN
+     * @param string|null $docCitation Citace
+     * @param string|null $docNote Poznámka k objednávce
+     * @param \DateTimeImmutable|null $dateRequested Requested date
      */
     public function __construct(
-        string $ticketDocDataSource,
-        string $eddSubtype,
-        string $docTitleIn,
-        string $docTitle,
-        ?string $documentId
+        public readonly TicketDataSource $ticketDocDataSource,
+        public readonly TicketEddSubtype $eddSubtype,
+        public readonly string $docTitleIn,
+        public readonly string $docTitle,
+        public readonly ?string $documentId = null,
+        public readonly ?array $documentAltIds = [],
+        public readonly ?string $docIdIn = null,
+        public readonly ?string $readerNote = null,
+        public readonly ?string $docNumberYear = null,
+        public readonly ?string $docNumberPyear = null,
+        public readonly ?string $docNumberPnumber = null,
+        public readonly ?string $docVolume = null,
+        public readonly ?int $pagesFrom = 0,
+        public readonly ?int $pagesTo = 0,
+        public readonly ?string $docAuthor = null,
+        public readonly ?string $docIssuer = null,
+        public readonly ?string $docISSN = null,
+        public readonly ?string $docISBN = null,
+        public readonly ?string $docCitation = null,
+        public readonly ?string $docNote = null,
+        public readonly ?DateTimeImmutable $dateRequested = null
     ) {
-        TicketEddDocDataSource::checkValue($ticketDocDataSource);
-        TicketEddSubtype::checkValue($eddSubtype);
-
-        if ($ticketDocDataSource === TicketEddDocDataSource::AUTO) {
-            if (!is_string($documentId)) {
-                throw new InvalidArgumentTypeException($documentId, 'string');
-            }
-        }
-
-        $this->ticketDocDataSource = $ticketDocDataSource;
-        $this->eddSubtype = $eddSubtype;
-        $this->docTitleIn = $docTitleIn;
-        $this->docTitle = $docTitle;
-        $this->documentId = $documentId;
     }
 
     /**
@@ -213,7 +72,7 @@ final class TicketEddRequest extends TicketRequest
     public function toArray(): array
     {
         $return = [
-            'ticket_type' => self::TICKET_TYPE,
+            'ticket_type' => self::TICKET_TYPE->value,
             'ticket_doc_data_source' => $this->ticketDocDataSource,
             'edd_subtype' => $this->eddSubtype,
             'doc_title_in' => $this->docTitleIn,
@@ -249,90 +108,5 @@ final class TicketEddRequest extends TicketRequest
         }
 
         return $return;
-    }
-
-    /**
-     * @param array<string> $documentAltIds
-     *
-     * @return void
-     */
-    public function setDocumentAltIds(array $documentAltIds): void
-    {
-        $this->documentAltIds = $documentAltIds;
-    }
-
-    public function setDocIdIn(?string $docIdIn): void
-    {
-        $this->docIdIn = $docIdIn;
-    }
-
-    public function setReaderNote(?string $readerNote): void
-    {
-        $this->readerNote = $readerNote;
-    }
-
-    public function setDocNumberYear(?string $docNumberYear): void
-    {
-        $this->docNumberYear = $docNumberYear;
-    }
-
-    public function setDocNumberPyear(?string $docNumberPyear): void
-    {
-        $this->docNumberPyear = $docNumberPyear;
-    }
-
-    public function setDocNumberPnumber(?string $docNumberPnumber): void
-    {
-        $this->docNumberPnumber = $docNumberPnumber;
-    }
-
-    public function setDocVolume(?string $docVolume): void
-    {
-        $this->docVolume = $docVolume;
-    }
-
-    public function setPagesFrom(?int $pagesFrom): void
-    {
-        $this->pagesFrom = $pagesFrom;
-    }
-
-    public function setPagesTo(?int $pagesTo): void
-    {
-        $this->pagesTo = $pagesTo;
-    }
-
-    public function setDocAuthor(?string $docAuthor): void
-    {
-        $this->docAuthor = $docAuthor;
-    }
-
-    public function setDocIssuer(?string $docIssuer): void
-    {
-        $this->docIssuer = $docIssuer;
-    }
-
-    public function setDocISSN(?string $docISSN): void
-    {
-        $this->docISSN = $docISSN;
-    }
-
-    public function setDocISBN(?string $docISBN): void
-    {
-        $this->docISBN = $docISBN;
-    }
-
-    public function setDocCitation(?string $docCitation): void
-    {
-        $this->docCitation = $docCitation;
-    }
-
-    public function setDocNote(?string $docNote): void
-    {
-        $this->docNote = $docNote;
-    }
-
-    public function setDateRequested(?DateTimeImmutable $dateRequested): void
-    {
-        $this->dateRequested = $dateRequested;
     }
 }
