@@ -11,30 +11,13 @@ use SmartEmailing\Types\StringType;
 final class Status
 {
     /**
-     * Status created datetime
-     *
-     * @var \DateTimeImmutable
+     * @param \DateTimeImmutable $createdAt Status created datetime
+     * @param \Mzk\ZiskejApi\Enum\StatusName $statusName Status name
      */
-    private DateTimeImmutable $createdAt;
-
-    /**
-     * Status name
-     *
-     * @var string
-     *
-     * @see \Mzk\ZiskejApi\Enum\StatusName
-     */
-    private string $name;
-
-    /**
-     * @throws \Consistence\Enum\InvalidEnumValueException
-     */
-    public function __construct(DateTimeImmutable $createdAt, string $name)
-    {
-        StatusName::checkValue($name);
-
-        $this->createdAt = $createdAt;
-        $this->name = $name;
+    public function __construct(
+        public readonly DateTimeImmutable $createdAt,
+        public readonly StatusName $statusName
+    ) {
     }
 
     /**
@@ -42,24 +25,13 @@ final class Status
      *
      * @return \Mzk\ZiskejApi\ResponseModel\Status
      *
-     * @throws \Consistence\Enum\InvalidEnumValueException
      * @throws \Exception
      */
     public static function fromArray(array $data): Status
     {
         return new self(
             new DateTimeImmutable(StringType::extract($data, 'date')),
-            StringType::extract($data, 'id')
+            StatusName::from(StringType::extract($data, 'id'))
         );
-    }
-
-    public function getCreatedAt(): DateTimeImmutable
-    {
-        return $this->createdAt;
-    }
-
-    public function getName(): string
-    {
-        return $this->name;
     }
 }
